Move source config bits into own module
